Fix test isolation: move compareDirectories tests to separate file
